py3: ported Fem to python3

+++ src/Mod/Fem/App/HypothesisPy.cpp
@@ -900,13 +900,13 @@ StdMeshers_NumberOfSegmentsPy::~StdMeshers_NumberOfSegmentsPy()
 
 Py::Object StdMeshers_NumberOfSegmentsPy::setNumSegm(const Py::Tuple& args)
 {
-    hypothesis<StdMeshers_NumberOfSegments>()->SetNumberOfSegments((int)Py::Int(args[0]));
+    hypothesis<StdMeshers_NumberOfSegments>()->SetNumberOfSegments((int)Py::Long(args[0]));
     return Py::None();
 }
 
 Py::Object StdMeshers_NumberOfSegmentsPy::getNumSegm(const Py::Tuple& args)
 {
-    return Py::Int(hypothesis<StdMeshers_NumberOfSegments>()->GetNumberOfSegments());
+    return Py::Long(hypothesis<StdMeshers_NumberOfSegments>()->GetNumberOfSegments());
 }
 
 // ----------------------------------------------------------------------------
@@ -931,13 +931,13 @@ StdMeshers_NumberOfLayersPy::~StdMeshers_NumberOfLayersPy()
 
 Py::Object StdMeshers_NumberOfLayersPy::setNumLayers(const Py::Tuple& args)
 {
-    hypothesis<StdMeshers_NumberOfLayers>()->SetNumberOfLayers((int)Py::Int(args[0]));
+    hypothesis<StdMeshers_NumberOfLayers>()->SetNumberOfLayers((int)Py::Long(args[0]));
     return Py::None();
 }
 
 Py::Object StdMeshers_NumberOfLayersPy::getNumLayers(const Py::Tuple& args)
 {
-    return Py::Int(hypothesis<StdMeshers_NumberOfLayers>()->GetNumberOfLayers());
+    return Py::Long(hypothesis<StdMeshers_NumberOfLayers>()->GetNumberOfLayers());
 }
 
 // ----------------------------------------------------------------------------
@@ -1011,13 +1011,13 @@ StdMeshers_LengthFromEdgesPy::~StdMeshers_LengthFromEdgesPy()
 
 Py::Object StdMeshers_LengthFromEdgesPy::setMode(const Py::Tuple& args)
 {
-    hypothesis<StdMeshers_LengthFromEdges>()->SetMode((int)Py::Int(args[0]));
+    hypothesis<StdMeshers_LengthFromEdges>()->SetMode((int)Py::Long(args[0]));
     return Py::None();
 }
 
 Py::Object StdMeshers_LengthFromEdgesPy::getMode(const Py::Tuple& args)
 {
-    return Py::Int(hypothesis<StdMeshers_LengthFromEdges>()->GetMode());
+    return Py::Long(hypothesis<StdMeshers_LengthFromEdges>()->GetMode());
 }
 
 // ----------------------------------------------------------------------------
